Remove SecurityManager singleton
	Removed singletone inheritance.
	Corrected usage of SecurityManager.
	Updated Unit Tests.

	Related to : APPLINK-21915

+++ src/components/config_profile/include/config_profile/profile.h
@@ -403,7 +403,7 @@ class Profile :public protocol_handler::ProtocolHandlerSettings,
   /**
    * @brief Return hours amount when PTU should be triggered
    */
-  uint32_t update_before_hours() const;
+  size_t update_before_hours() const;
 
 #endif //ENABLE_SECURITY
     /**
